+++ src/main/java/uk/gov/hmcts/darts/authorisation/component/impl/CaseIdControllerAuthorisationImpl.java
@@ -1,0 +1,77 @@
+package uk.gov.hmcts.darts.authorisation.component.impl;
+
+import com.fasterxml.jackson.databind.JsonNode;
+import jakarta.servlet.http.HttpServletRequest;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Component;
+import uk.gov.hmcts.darts.authorisation.component.Authorisation;
+import uk.gov.hmcts.darts.authorisation.component.ControllerAuthorisation;
+import uk.gov.hmcts.darts.authorisation.enums.ContextIdEnum;
+import uk.gov.hmcts.darts.common.enums.SecurityRoleEnum;
+import uk.gov.hmcts.darts.common.exception.DartsApiException;
+
+import java.util.Optional;
+import java.util.Set;
+
+import static uk.gov.hmcts.darts.authorisation.enums.ContextIdEnum.CASE_ID;
+import static uk.gov.hmcts.darts.authorisation.exception.AuthorisationError.BAD_REQUEST_CASE_ID;
+
+@Component
+@RequiredArgsConstructor
+@Slf4j
+class CaseIdControllerAuthorisationImpl extends BaseControllerAuthorisation
+    implements ControllerAuthorisation {
+
+    private static final String CASE_ID_PARAM = "case_id";
+
+    private final Authorisation authorisation;
+
+    @Override
+    public ContextIdEnum getContextId() {
+        return CASE_ID;
+    }
+
+    @Override
+    public void checkAuthorisation(HttpServletRequest request, Set<SecurityRoleEnum> roles) {
+        Optional<String> caseIdParamOptional = getPathParamValue(request, CASE_ID_PARAM);
+        checkAuthorisationByCaseId(caseIdParamOptional, roles);
+
+        if (caseIdParamOptional.isEmpty()) {
+            caseIdParamOptional = Optional.ofNullable(request.getParameter(CASE_ID_PARAM));
+            checkAuthorisationByCaseId(caseIdParamOptional, roles);
+        }
+
+        if (caseIdParamOptional.isEmpty()) {
+            caseIdParamOptional = Optional.ofNullable(request.getHeader(CASE_ID_PARAM));
+            checkAuthorisationByCaseId(caseIdParamOptional, roles);
+        }
+
+        if (caseIdParamOptional.isEmpty()) {
+            log.error(String.format(
+                BAD_REQUEST_AUTHORISATION_PARAM_ERROR_MESSAGE,
+                CASE_ID_PARAM,
+                request.getRequestURI()
+            ));
+            throw new DartsApiException(BAD_REQUEST_CASE_ID);
+        }
+    }
+
+    @Override
+    public void checkAuthorisation(JsonNode jsonNode, Set<SecurityRoleEnum> roles) {
+        authorisation.authoriseByCaseId(jsonNode.path(CASE_ID_PARAM).intValue(), roles);
+    }
+
+    private void checkAuthorisationByCaseId(Optional<String> caseIdParamOptional, Set<SecurityRoleEnum> roles) {
+        if (caseIdParamOptional.isPresent()) {
+            try {
+                Integer caseId = Integer.valueOf(caseIdParamOptional.get());
+                authorisation.authoriseByCaseId(caseId, roles);
+            } catch (NumberFormatException e) {
+                log.error("Unable to parse case_id for checkAuthorisation", e);
+                throw new DartsApiException(BAD_REQUEST_CASE_ID);
+            }
+        }
+    }
+
+}
