+++ src/main/java/uk/gov/hmcts/darts/authorisation/component/impl/TranscriptionIdControllerAuthorisationImpl.java
@@ -1,0 +1,78 @@
+package uk.gov.hmcts.darts.authorisation.component.impl;
+
+import com.fasterxml.jackson.databind.JsonNode;
+import jakarta.servlet.http.HttpServletRequest;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Component;
+import uk.gov.hmcts.darts.authorisation.component.Authorisation;
+import uk.gov.hmcts.darts.authorisation.component.ControllerAuthorisation;
+import uk.gov.hmcts.darts.authorisation.enums.ContextIdEnum;
+import uk.gov.hmcts.darts.common.enums.SecurityRoleEnum;
+import uk.gov.hmcts.darts.common.exception.DartsApiException;
+
+import java.util.Optional;
+import java.util.Set;
+
+import static uk.gov.hmcts.darts.authorisation.enums.ContextIdEnum.TRANSCRIPTION_ID;
+import static uk.gov.hmcts.darts.authorisation.exception.AuthorisationError.BAD_REQUEST_TRANSCRIPTION_ID;
+
+@Component
+@RequiredArgsConstructor
+@Slf4j
+class TranscriptionIdControllerAuthorisationImpl extends BaseControllerAuthorisation
+    implements ControllerAuthorisation {
+
+    private static final String TRANSCRIPTION_ID_PARAM = "transcription_id";
+
+    private final Authorisation authorisation;
+
+    @Override
+    public ContextIdEnum getContextId() {
+        return TRANSCRIPTION_ID;
+    }
+
+    @Override
+    public void checkAuthorisation(HttpServletRequest request, Set<SecurityRoleEnum> roles) {
+        Optional<String> transcriptionIdParamOptional = getPathParamValue(request, TRANSCRIPTION_ID_PARAM);
+        checkAuthorisationByTranscriptionId(transcriptionIdParamOptional, roles);
+
+        if (transcriptionIdParamOptional.isEmpty()) {
+            transcriptionIdParamOptional = Optional.ofNullable(request.getParameter(TRANSCRIPTION_ID_PARAM));
+            checkAuthorisationByTranscriptionId(transcriptionIdParamOptional, roles);
+        }
+
+        if (transcriptionIdParamOptional.isEmpty()) {
+            transcriptionIdParamOptional = Optional.ofNullable(request.getHeader(TRANSCRIPTION_ID_PARAM));
+            checkAuthorisationByTranscriptionId(transcriptionIdParamOptional, roles);
+        }
+
+        if (transcriptionIdParamOptional.isEmpty()) {
+            log.error(String.format(
+                BAD_REQUEST_AUTHORISATION_PARAM_ERROR_MESSAGE,
+                TRANSCRIPTION_ID_PARAM,
+                request.getRequestURI()
+            ));
+            throw new DartsApiException(BAD_REQUEST_TRANSCRIPTION_ID);
+        }
+    }
+
+    @Override
+    public void checkAuthorisation(JsonNode jsonNode, Set<SecurityRoleEnum> roles) {
+        authorisation.authoriseByTranscriptionId(jsonNode.path(TRANSCRIPTION_ID_PARAM).intValue(), roles);
+    }
+
+    private void checkAuthorisationByTranscriptionId(Optional<String> transcriptionIdParamOptional,
+                                                     Set<SecurityRoleEnum> roles) {
+        if (transcriptionIdParamOptional.isPresent()) {
+            try {
+                Integer transcriptionId = Integer.valueOf(transcriptionIdParamOptional.get());
+                authorisation.authoriseByTranscriptionId(transcriptionId, roles);
+            } catch (NumberFormatException e) {
+                log.error("Unable to parse transcription_id for checkAuthorisation", e);
+                throw new DartsApiException(BAD_REQUEST_TRANSCRIPTION_ID);
+            }
+        }
+    }
+
+}
