+++ src/main/java/uk/gov/hmcts/darts/authorisation/service/ControllerAuthorisationFactory.java
@@ -1,0 +1,10 @@
+package uk.gov.hmcts.darts.authorisation.service;
+
+import uk.gov.hmcts.darts.authorisation.component.ControllerAuthorisation;
+import uk.gov.hmcts.darts.authorisation.enums.ContextIdEnum;
+
+public interface ControllerAuthorisationFactory {
+
+    ControllerAuthorisation getHandler(final ContextIdEnum contextId);
+
+}
